PERA-1381 :: remove remote config changes
