+++ common-sdk/src/main/kotlin/com/algorand/wallet/account/custom/data/mapper/model/CustomInfoMapperImpl.kt
@@ -13,7 +13,6 @@
 package com.algorand.wallet.account.custom.data.mapper.model
 
 import com.algorand.wallet.account.custom.data.database.model.CustomInfoEntity
-import com.algorand.wallet.account.custom.domain.mapper.model.CustomInfoMapper
 import com.algorand.wallet.account.custom.domain.model.CustomInfo
 import javax.inject.Inject
