--- common-sdk/src/test/kotlin/com/algorand/wallet/remoteconfig/repository/FeatureToggleRepositoryImplTest.kt
@@ -1,59 +1,0 @@
-/*
- * Copyright 2022 Pera Wallet, LDA
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License
- */
-
-package com.algorand.wallet.remoteconfig.data.repository
-
-import com.algorand.wallet.remoteconfig.domain.service.FirebaseRemoteConfigService
-import io.mockk.coEvery
-import io.mockk.every
-import io.mockk.mockk
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertFalse
-import org.junit.Assert.assertTrue
-import org.junit.Test
-import kotlinx.coroutines.test.runTest
-
-class FeatureToggleRepositoryImplTest {
-
-    private val firebaseRemoteConfigService: FirebaseRemoteConfigService = mockk()
-    private val sut = FeatureToggleRepositoryImpl(firebaseRemoteConfigService)
-
-    @Test
-    fun `EXPECT successful completion WHEN initializing operational toggles`() = runTest {
-        coEvery { firebaseRemoteConfigService.fetchRemoteConfig() } returns Unit
-
-        val result = sut.initializeOperationalToggles()
-
-        assertEquals(Unit, result)
-    }
-
-    @Test
-    fun `EXPECT true WHEN checking enabled feature`() = runTest {
-        val featureName = "enabledFeature"
-        every { firebaseRemoteConfigService.getBoolean(featureName) } returns true
-
-        val result = sut.isFeatureEnabled(featureName)
-
-        assertTrue(result)
-    }
-
-    @Test
-    fun `EXPECT false WHEN checking disabled feature`() = runTest {
-        val featureName = "disabledFeature"
-        every { firebaseRemoteConfigService.getBoolean(featureName) } returns false
-
-        val result = sut.isFeatureEnabled(featureName)
-
-        assertFalse(result)
-    }
-}
-
